Add missing image types to image tests
