+++ utils/test_harness/include/test_harness/test_harness_image.hpp
@@ -18,12 +18,6 @@
 namespace lzt = level_zero_tests;
 namespace level_zero_tests {
 
-struct Dims {
-  uint64_t width;
-  uint32_t height;
-  uint32_t depth;
-};
-
 const ze_image_flags_t image_rw_flags[2] = {0, ZE_IMAGE_FLAG_KERNEL_WRITE};
 const ze_image_flags_t image_cache_flags[2] = {0, ZE_IMAGE_FLAG_BIAS_UNCACHED};
 
@@ -96,12 +90,6 @@ const std::vector<ze_image_format_swizzle_t> image_format_swizzles_all = {
     ZE_IMAGE_FORMAT_SWIZZLE_X};
 
 size_t get_format_component_count(ze_image_format_layout_t layout);
-std::string shortened_string(ze_image_type_t type);
-Dims get_sample_image_dims(ze_image_type_t image_type);
-std::vector<ze_image_type_t> get_supported_image_types(
-    ze_device_handle_t device,
-    bool exclude_arrays = false,
-    bool exclude_buffer = false);
 bool image_support();
 void print_image_format_descriptor(const ze_image_format_t descriptor);
 void print_image_descriptor(const ze_image_desc_t descriptor);
